backPlaytherapy/requirements.txt
Django==2.2.5
django-appconf==1.0.3
django-cors-headers==3.1.0
django-select2==7.1.1
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
gunicorn==20.0.4
psycopg2==2.8.3
PyJWT==1.7.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==5.1.0
